ansible/files/radem-notebooks/pyproject.toml
[tool.poetry]
name = "notebooks"
version = "0.1.0"
description = ""
authors = ["Krzysztof Pęczek <krpeczek@student.agh.edu.pl>", "Szymon Bednorz <sbednorz@student.agh.edu.pl>"]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.1"
seaborn = "^0.13.2"
matplotlib = "^3.8.3"
spacepy = "^0.5.0"
plotly = "^5.19.0"
ipympl = "^0.9.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
